Fixed project file structure, brought contents of BlueConnect to root directory.
