--- BlueConnect/app/src/main/res/layout/activity_main.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/listDevices"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/white"
-    tools:context=".MainActivity">
-
-    <Button
-        android:id="@+id/listen"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="30dp"
-        android:layout_marginTop="42dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginBottom="18dp"
-        android:text="Listen"
-        app:layout_constraintBottom_toTopOf="@+id/listview"
-        app:layout_constraintEnd_toStartOf="@+id/listdevices"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/listdevices"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="42dp"
-        android:layout_marginEnd="48dp"
-        android:layout_marginBottom="18dp"
-        android:text="List Devices"
-        app:layout_constraintBottom_toTopOf="@+id/listview"
-        app:layout_constraintEnd_toStartOf="@+id/status"
-        app:layout_constraintStart_toEndOf="@+id/listen"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ListView
-        android:id="@+id/listview"
-        android:layout_width="407dp"
-        android:layout_height="272dp"
-        android:layout_marginStart="3dp"
-        android:layout_marginEnd="1dp"
-        android:layout_marginBottom="39dp"
-        android:focusable="true"
-        android:clickable="true"
-        app:layout_constraintBottom_toTopOf="@+id/msg"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/msg"
-        android:layout_width="396dp"
-        android:layout_height="132dp"
-        android:layout_marginStart="7dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="94dp"
-        android:text="Message"
-        app:layout_constraintBottom_toTopOf="@+id/writemsg"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/status"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="54dp"
-        android:layout_marginEnd="36dp"
-        android:layout_marginBottom="35dp"
-        android:text="Status"
-        app:layout_constraintBottom_toTopOf="@+id/listview"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <EditText
-        android:id="@+id/writemsg"
-        android:layout_width="218dp"
-        android:layout_height="54dp"
-        android:layout_marginStart="24dp"
-        android:layout_marginBottom="27dp"
-        android:text="Message"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/send"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="45dp"
-        android:layout_marginTop="95dp"
-        android:layout_marginEnd="36dp"
-        android:layout_marginBottom="32dp"
-        android:text="Send"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/writemsg"
-        app:layout_constraintTop_toBottomOf="@+id/msg" />
-</androidx.constraintlayout.widget.ConstraintLayout>
