--- BlueConnect/build.gradle.kts
@@ -1,4 +1,0 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    id("com.android.application") version "8.1.1" apply false
-}
