--- BlueConnect/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Sep 21 12:52:35 PDT 2023
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.0-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
