+++ local.properties
@@ -7,4 +7,4 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-sdk.dir=C\:\\Users\\Nick\\AppData\\Local\\Android\\Sdk
+sdk.dir=/home/chris/Android/Sdk
