--- phase1/README.md
@@ -1,21 +1,0 @@
-# Phase I: Analyzing Users, Competitors, and Initial Designs
-
-## Introduction
-
-  Our innovative application, BlueConnect, is driven by a mission to empower users with the seamless capability to establish connections with multiple Bluetooth devices simultaneously. What sets BlueConnect apart from the competition is its unparalleled versatility. Unlike other solutions on the market, BlueConnect transcends brand limitations, allowing users to seamlessly connect Bluetooth devices regardless of their make or model. This means that whether you're pairing smartphones, tablets, headphones, speakers, or any other Bluetooth-enabled gadget, BlueConnect delivers a universal and hassle-free connection experience. In conclusion, BlueConnect isn't just a Bluetooth connectivity solution; it's a game-changer that empowers users to navigate the complexities of the modern tech landscape effortlessly. We invite you to experience the freedom, convenience, and reliability that BlueConnect offers as you connect with multiple Bluetooth devices, unbounded by brand constraints.
-
-## Methods
-
-  We utilized Personas/Scenarios to gain new insight into our users needs and potential UX problems while Sketches were used to get the flow of the program. The purpose of using personas and scenarios in our research was to develop a deeper understanding of our target audience or users and to identify potential main points, needs, or opportunities related to our BlueConnect application. The first method we used was personas which helped us to create detailed and realistic user personas representing different segments of our target audience. The second method was scenarios which helped us to create realistic scenarios or use cases that reflect how users interact with our product.User testing was also used, as our developers would use our application to both make sure it is working appropriately and use it as a user would, also to validate our personas and scenarios and gain insights into how users would interact with your product or service. Sketches were used to take our ideas on the application and turn it into how the application would run. The sketches were used to get the baseline of how the application would work: a user connects one or more devices and can adjust the settings of said device(s).
-
-## Findings
-
-  With establishing how BlueConnect was to achieve the goals for a given persona (and scenarios that a persona would be in) and how the program was to flow through the sketches/diagrams, there were changes made in how BlueConnect was to be implemented and approached during this phase. The focus of this phase was to establish a connection for a given persona. The application at the end of this phase was able to get a person connected but there was still a concern for identifying a device. A given UUID had to be hardcoded to identify a device and this is going to be the main focus for Phase 2 - how to get multiple connections for a user. The sketches created during the initial brainstorming were linear in how the application would open a different screen in response to pressing a button. With the first phase over, most of the predetermined functions are now part of the home screen. A user can view a list of devices to connect to, click the “Listen” button to get a connection, and view the status of a given connection. During implementation, a new perspective was gained with a chat functionality added to the bottom of the home screen. The addition of messaging and having most of the functions in the home screen are welcome changes and will be used to create a new approach to how the application will flow for future diagrams.
-
-## Conclusions
-
-  Using personas and scenarios helped us to determine what specific user needs our project should focus on addressing, as well as what potential difficulties users could encounter while using our product.  It has not been implemented yet, but our findings indicated that it would be helpful to the user to separate device type and input/output to eliminate possible confusion. Sketches were able to show the foundation of the application - connecting to a device - but there are other facets that haven’t been thought of… Chat functionality created during implementation shows there are more ideas that can be used and considered for this project.
-
-## Caveats
-
-  While personas and scenarios are useful, they cannot perfectly replicate the experience of a real user utilizing our product.  Additionally, the user testing was done by team members who were already familiar with the application and how it worked, so it is difficult to gain new insight from it. There was a limitation to the amount of devices to test, but a work around would be to emulate an Android phone.
